[poincare/n_ary_expression] Approx children to double to know if they are integers
Fixes "gcd(1-e^(-90), 0)" that was giving a result while it's undef
Fixes inconsistencies that created a crash in additionnal results of
"sin(gcd(1-e^(-90), 0))" because of the same reason

+++ poincare/src/n_ary_expression.cpp
@@ -102,7 +102,7 @@ Expression NAryExpression::checkChildrenAreRationalIntegersAndUpdate(
         return replaceWithUndefinedInPlace();
       }
       // If c was complex but with a null imaginary part, real part is checked.
-      float app = c.approximateToScalar<float>(
+      double app = c.approximateToScalar<double>(
           reductionContext.context(), reductionContext.complexFormat(),
           reductionContext.angleUnit(), true);
       if (std::isfinite(app) && app != std::round(app)) {
